execution/mt5_executor.py
import MetaTrader5 as mt5
from config import SYMBOL

def connect_mt5():
    if not mt5.initialize():
        raise RuntimeError("MT5 not initialized")

def place_trade(direction, lot, sl, tp):
    price = mt5.symbol_info_tick(SYMBOL).ask if direction == "BUY" else mt5.symbol_info_tick(SYMBOL).bid
    request = {
        "action": mt5.TRADE_ACTION_DEAL,
        "symbol": SYMBOL,
        "volume": lot,
        "type": mt5.ORDER_TYPE_BUY if direction == "BUY" else mt5.ORDER_TYPE_SELL,
        "price": price,
        "sl": sl,
        "tp": tp,
        "deviation": 20,
        "magic": 26062026,
        "comment": "ICT_AI_MENTOR",
        "type_time": mt5.ORDER_TIME_GTC,
        "type_filling": mt5.ORDER_FILLING_IOC,
    }
    result = mt5.order_send(request)
    return result
