requirements.txt
pandas
pytz
# requirements.txt for AstroQuant Day Planner
PyYAML
yfinance
flatlib
feedparser
pandas
requests
